ColorRGB moved to a separate file
